+++ logging_config_test.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"os"
 	"testing"
 
 	mockassert "github.com/derision-test/go-mockgen/testutil/assert"
@@ -9,47 +10,39 @@ import (
 )
 
 func TestLoggingConfigLoadLogs(t *testing.T) {
+	os.Setenv("X", "foo")
+	os.Setenv("Y", "123")
+	os.Setenv("W", `["bar", "baz", "bonk"]`)
+
 	type C struct {
 		X string   `env:"x"`
 		Y int      `env:"y"`
 		Z []string `env:"w" display:"Q"`
 	}
 
-	config := NewMockConfig()
 	logger := NewMockLogger()
-	lc := NewLoggingConfig(config, logger, nil)
+	lc := NewConfig(NewEnvSourcer(""), WithLogger(logger))
+
 	chunk := &C{}
-
-	config.LoadFunc.SetDefaultHook(func(target interface{}, modifiers ...TagModifier) error {
-		target.(*C).X = "foo"
-		target.(*C).Y = 123
-		target.(*C).Z = []string{"bar", "baz", "bonk"}
-		return nil
-	})
-
 	require.Nil(t, lc.Load(chunk))
 	mockassert.CalledOnceWith(t, logger.PrintfFunc, mockassert.Values("Config loaded: %s", "\nQ=[\"bar\",\"baz\",\"bonk\"]\nX=foo\nY=123"))
 }
 
 func TestLoggingConfigMask(t *testing.T) {
+	os.Setenv("X", "foo")
+	os.Setenv("Y", "123")
+	os.Setenv("W", `["bar", "baz", "bonk"]`)
+
 	type C struct {
 		X string   `env:"x"`
 		Y int      `env:"y" mask:"true"`
 		Z []string `env:"w" mask:"true"`
 	}
 
-	config := NewMockConfig()
 	logger := NewMockLogger()
-	lc := NewLoggingConfig(config, logger, nil)
+	lc := NewConfig(NewEnvSourcer(""), WithLogger(logger))
+
 	chunk := &C{}
-
-	config.LoadFunc.SetDefaultHook(func(target interface{}, modifiers ...TagModifier) error {
-		target.(*C).X = "foo"
-		target.(*C).Y = 123
-		target.(*C).Z = []string{"bar", "baz", "bonk"}
-		return nil
-	})
-
 	require.Nil(t, lc.Load(chunk))
 	mockassert.CalledOnceWith(t, logger.PrintfFunc, mockassert.Values("Config loaded: %s", "\nX=foo"))
 }
@@ -59,10 +52,9 @@ func TestLoggingConfigBadMaskTag(t *testing.T) {
 		X string `env:"x" mask:"34"`
 	}
 
-	config := NewMockConfig()
 	logger := NewMockLogger()
-	lc := NewLoggingConfig(config, logger, nil)
-	chunk := &C{}
+	lc := NewConfig(NewEnvSourcer(""), WithLogger(logger))
 
+	chunk := &C{}
 	assert.EqualError(t, lc.Load(chunk), "failed to serialize config (field 'X' has an invalid mask tag)")
 }